requirements.txt
fastapi
uvicorn
groq
pydantic
python-dotenv
httpx
tavily-python
